Convert from ts to js
