--- utils/supabase/component.ts
@@ -1,10 +1,0 @@
-import { createBrowserClient } from '@supabase/ssr'
-
-export function createClient() {
-  const supabase = createBrowserClient(
-    process.env.NEXT_PUBLIC_SUPABASE_URL!,
-    process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
-  )
-
-  return supabase
-}
