+++ utils/supabase/static-props.js
@@ -1,0 +1,14 @@
+import { createClient as createClientPrimitive } from '@supabase/supabase-js';
+
+export function createClient() {
+  const url = process.env.NEXT_PUBLIC_SUPABASE_URL;
+  const anonKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY;
+
+  if (!url || !anonKey) {
+    throw new Error('Missing Supabase environment variables');
+  }
+
+  const supabase = createClientPrimitive(url, anonKey);
+
+  return supabase;
+}
